Flocking not working as well as intended but great jellyfish animation methods
Jellyfish velocity is always in the Up vector which keeps them inline but it's not working with for flocking. I'm using an oscillating animation which slow and  speeds up based on the animation of the jellyfish. I'm going to change some of this to make them better cohese but I'm liking the motion as it as aside from them not flocking together well.